An example utilized the first several functions
This is an example that I applied the functions defined in the utils.py and functions.py (master branch) to the web crawling. As you can see, it was extremely fixed. That was really annoying because we had to make modifications constantly. It  was far from an easily appliable template, and that's why we soon abandoned those functions and figured out new ones.