A complete example
This is a complete example utilizing the functions created in utils.py and functions.py. I experimented on Guangdong Provincial Website, simply because we were working in the capital city of Guangdong Province, Guangzhou. Actually our method was effective, successfully crawling the data in most pages. It didn't mean our codes were perfect though.

 As you can see, the codes are too redundant,  due to our procedure-oriented likeliness. It is because our program is not object-oriented enough that we could not deal with the extreme situations in Python. Rather we naively categorized the problematic pages, manually copied down the information to our Excel file.

Funny, isn't it? From then on, we were more and more focused on what kinds of pages the codes didn't work, which from my point of view is really helpful to our progress.